imgpkg seems to try HTTP first with registries...
By default, the Docker registry will serve itself over HTTPS. However,
its web server is not configured to redirect HTTP requests to HTTPS.

This is a problem with imgpkg, as it uses Go's `net/http` RoundTripper
to connect to the registry and assumes a HTTP connection under some
circumstances.

This commit fronts the Docker registry with nginx so that it can handle
HTTP to HTTPS rewrites.

Signed-off-by: Carlos Nunez <ncarlos@vmware.com>